Download Documents

Download Documents

hal_download(x, ...)

# S3 method for HALQuery
hal_download(
  x,
  limit = 30,
  start = 0,
  progress = getOption("odyssey.progress"),
  verbose = getOption("odyssey.verbose"),
  ...
)

Arguments

x

An object of class HALQuery (typically returned by hal_api()).

...

Currently not used.

limit

An integer giving the maximum number of results. According to HAL policy, it cannot exceed 10000.

start

An integer specifying an absolute offset in the complete sorted list of matches to be used as the beginning of the current page. Only used if cursor is FALSE.

progress

A logical scalar: should a progress bar for for the request be printed? Only used if R is being used interactively.

verbose

A logical scalar: should extra information be reported?

Value

An (invisible) logical scalar, TRUE for success and FALSE if any failure.

See also

Author

N. Frerebeau

Examples

if (FALSE) {
library(magrittr)

## Download the 10 most recent archaeological publication
## (if any files)
hal_api() %>%
  hal_query("archéologie") %>%
  hal_filter("ART" %IN% "docType_s") %>%
  hal_sort("producedDate_tdate", decreasing = TRUE) %>%
  hal_download(limit = 10)
}