Filter Results

Filter Results

hal_filter(x, ...)

# S3 method for HALQuery
hal_filter(x, value, field = NULL, ...)

Arguments

x
An object of class HALQuery (typically returned by hal_api).

...
Currently not used.

value
A character string specifying the value to be used to filter the results.

field
A character string specifying the field to filter along.

Value

An object of class HALQuery.

See also

Author

N. Frerebeau

Examples

library(magrittr)

## Simple filer
hal1 <- hal_api() %>% hal_filter("submitType_s", "file")
hal1$fq

#> [1] "file:submitType_s"

## Advanced filter
hal2 <- hal_api() %>% hal_filter("docType_s", c("THESE", "HDR"))
hal2$fq

#> [1] "THESE:docType_s" "HDR:docType_s"  

## Multiple filters
hal3 <- hal_api() %>%
  hal_filter("submittedDate_tdate", "[NOW-1MONTHS/DAY TO NOW/HOUR]") %>%
  hal_filter("submitType_s", "-notice")
hal3$fq

#> [1] "[NOW-1MONTHS/DAY TO NOW/HOUR]:submittedDate_tdate"
#> [2] "-notice:submitType_s"                             

## Range filters
hal4 <- hal_api() %>% hal_filter("submittedDateY_i", "[2000 TO 2013]")
hal4$fq

#> [1] "[2000 TO 2013]:submittedDateY_i"

hal5 <- hal_api() %>% hal_filter("city_s", "[Aa TO Ab]")
hal5$fq

#> [1] "[Aa TO Ab]:city_s"