Group Results

Group Results

hal_group(x, ...)

# S3 method for HALQuery
hal_group(x, field, limit = 1, sort = "score", decreasing = FALSE, ...)

Arguments

x
An object of class HALQuery (typically returned by hal_api).

...
Currently not used.

field
A character string specifying the field to group by.

limit
An integer giving the maximum number of results per group.

sort
A character string specifying the field to be used to sort the results.

decreasing
A logical scalar: should the sort be increasing or decreasing?

Value

An object of class HALQuery.

See also

Author

N. Frerebeau

Examples

# \donttest{
library(magrittr)

## Most recent publication by journal
hal_api() %>%
  hal_query("archéologie") %>%
  hal_select("producedDate_tdate") %>%
  hal_filter("docType_s" %IN% "ART") %>%
  hal_sort("producedDate_tdate", decreasing = TRUE) %>%
  hal_group(field = "journalTitle_s", limit = 1,
            sort = "producedDate_tdate", decreasing = TRUE) %>%
  hal_search(limit = 10)

#> 

#>                                                                          groupValue
#> 1                                                          Quaternary Geochronology
#> 2                                                             Nature Communications
#> 3                                           International Journal of Paleopathology
#> 4                                                                Scientific Reports
#> 5                                                        L'information grammaticale
#> 6                                                                      Archaeometry
#> 7  Annuaire de l’École pratique des hautes études. Section des sciences religieuses
#> 8                                                   Revue archéologique de Picardie
#> 9                                                                Métropolitiques.eu
#> 10                                    Journal of Data Mining and Digital Humanities
#>    numFound start   producedDate_tdate
#> 1        30     0 2022-01-01T00:00:00Z
#> 2         3     0 2021-12-01T00:00:00Z
#> 3         6     0 2021-12-01T00:00:00Z
#> 4        20     0 2021-12-01T00:00:00Z
#> 5         1     0 2021-10-01T00:00:00Z
#> 6        35     0 2021-10-01T00:00:00Z
#> 7        17     0 2021-10-01T00:00:00Z
#> 8        58     0 2021-10-01T00:00:00Z
#> 9         1     0 2021-09-30T00:00:00Z
#> 10        4     0 2021-09-28T00:00:00Z

# }