Search

Search

hal_count(x, ...)

hal_search(x, ...)

# S3 method for HALQuery
hal_count(x, path = "search", ...)

# S3 method for HALQuery
hal_search(
  x,
  path = "search",
  limit = 30,
  start = 0,
  cursor = FALSE,
  parse = "df",
  concat = ",",
  progress = getOption("odyssey.progress"),
  verbose = getOption("odyssey.verbose"),
  ...
)

Arguments

x
An object of class HALQuery (typically returned by hal_api).

...
Currently not used.

path
A character string specifying the url path (see details).

limit
An integer giving the maximum number of results. According to HAL policy, it cannot exceed 10000.

start
An integer specifying an absolute offset in the complete sorted list of matches to be used as the beginning of the current page. Only used if cursor is FALSE.

cursor
A logical scalar: should a cursor be used for the pagination of results? If TRUE, the sort parameter of the query will set to "docid asc".

parse
A character string specifying the type of the results. It must be one of "df" (default) or "list" (see SolrClient).

concat
A character string specifying the character to concatenate elements of longer than length \(1\).

progress
A logical scalar: should a progress bar for for the request be printed? Only used if R is being used interactively.

verbose
A logical scalar: should extra information be reported?

Value

hal_count returns the total number of results.

hal_search returns a data.frame or list (according to type).

References

Apache Solr documentation.

HAL search documentation.

HAL reference frame documentation.

See also

Author

N. Frerebeau

Examples

# \donttest{
library(magrittr)

## Simple search
topic <- list("archéologie", "Celtes", "France") # Combined with AND
hal_api() %>%
  hal_query(topic) %>%
  hal_search(limit = 10)

#> 10 documents out of a maximum of 74 were returned.

#> # A tibble: 10 x 2
#>      docid label_s                                                              
#>      <int> <chr>                                                                
#>  1  106235 Germaine Leman-Delerive. La Belgique et le Nord de la France du IIIè…
#>  2  106209 Germaine Leman-Delerive, Eugène Warmembol. La Belgique et le Nord de…
#>  3  818910 Thierry Janin, Michel Py. Grecs et Celtes en Languedoc. Hermary A. e…
#>  4  962598 S. Chevrier, Stephane Verger. La complémentarité entre les dépôts de…
#>  5 2031928 Rapport d’évaluation d’une entité de recherche. AOROC - Archéologie …
#>  6  962465 Stephane Verger. Trésors de femmes en France centrale et méridionale…
#>  7 1094643 Thierry Lejars. Images celtiques du Pays de France. Gaulois d'ici & …
#>  8 1094653 Thierry Lejars. Le cimetière de la Fosse-Cotheret à Roissy-en-France…
#>  9  962398 Stephane Verger. Les femmes du Premier âge du Fer en France centrale…
#> 10  963605 Stephane Verger. Avant propos. Sophie Krausz, Anne Colin, Katherine …

## Get a list of archaeological journals
topic <- c("archéologie", "archaeology", "archäologie") # Combined with OR
hal_api() %>%
  hal_query(topic) %>%
  hal_select("title_s", "issn_s") %>%
  hal_filter("" %TO% "*" %IN% "issn_s") %>%
  hal_sort("title_s") %>%
  hal_search(path = "ref/journal")

#> 30 documents out of a maximum of 218 were returned.

#> # A tibble: 30 x 2
#>    title_s                                                            issn_s   
#>    <chr>                                                              <chr>    
#>  1 ACUA. Underwater Archaeology Proceedings                           1074-3421
#>  2 ADLFI. Archéologie de la France - Informations [En ligne]          2114-0502
#>  3 AIMA Newsletter                                                    1446-8948
#>  4 Afrique : archéologie et arts                                      1634-3123
#>  5 Afrique: Archéologie & Arts                                        2431-2045
#>  6 Agri centuriati. An International Journal of Landscape Archaeology 1825-1277
#>  7 Americae. European Journal of Americanist Archaeology              2497-1510
#>  8 American Antiquity                                                 0002-7316
#>  9 American Journal of Archaeology                                    0002-9114
#> 10 Anglo-Saxon Studies in Archaeology and History                     02645254 
#> # … with 20 more rows

## Get the most recent archaeological publication (in French) by journal
hal_api() %>%
  hal_query("archéologie") %>%
  hal_select("producedDate_tdate") %>%
  hal_filter("ART" %IN% "docType_s") %>%
  hal_sort("producedDate_tdate", decreasing = TRUE) %>%
  hal_group(
    field = "journalTitle_s",
    sort = "producedDate_tdate",
    decreasing = TRUE
  ) %>%
  hal_search(limit = 10)

#> 

#>                                        groupValue numFound start
#> 1                             Ardèche archéologie       52     0
#> 2               Journal of African Earth Sciences        1     0
#> 3  Bulletin de la Société française d'égyptologie       20     0
#> 4      Journal of Archaeological Science: Reports       72     0
#> 5                              Scientific Reports       13     0
#> 6                                Heritage Science        7     0
#> 7    Physics of the Earth and Planetary Interiors        6     0
#> 8          Journal of Anthropological Archaeology       11     0
#> 9   Journal of Data Mining and Digital Humanities        3     0
#> 10       Geoarchaeology: An International Journal        6     0
#>      producedDate_tdate
#> 1  2021-01-01T00:00:00Z
#> 2  2020-12-01T00:00:00Z
#> 3  2020-12-01T00:00:00Z
#> 4  2020-12-01T00:00:00Z
#> 5  2020-12-01T00:00:00Z
#> 6  2020-12-01T00:00:00Z
#> 7  2020-12-01T00:00:00Z
#> 8  2020-12-01T00:00:00Z
#> 9  2020-11-14T00:00:00Z
#> 10 2020-11-12T00:00:00Z

## Get a list of archaeological laboratories
## (only joint laboratories of the CNRS and a French university)
topic <- "archéologie" %IN% "text" %AND% "UMR" %IN% "code_t"
hal_api() %>%
  hal_query(topic) %>%
  hal_select("name_s", "acronym_s", "code_s") %>%
  hal_filter("VALID" %IN% "valid_s") %>%
  hal_sort("acronym_s", decreasing = TRUE) %>%
  hal_search(path = "ref/structure", limit = 15)

#> 15 documents out of a maximum of 17 were returned.

#> # A tibble: 15 x 3
#>    name_s                                             acronym_s code_s          
#>    <chr>                                              <chr>     <chr>           
#>  1 ORIENT ET MÉDITERRANÉE : Textes, Archéologie, His… OM        UMR8167         
#>  2 Laboratoire Archéologie et Territoires             LAT       UMR6575         
#>  3 Laboratoire d'Archéologie Moléculaire et Structur… LAMS      UMR8220         
#>  4 Laboratoire de Médiévistique Occidentale de Paris  LAMOP     UMR8589         
#>  5 Laboratoire d'Archéologie Médiévale et Moderne en… LA3M      UMR7298         
#>  6 Histoire Archéologie Littérature des Mondes Ancie… HALMA     UMR8164         
#>  7 Centre de Recherche en Archéologie, Archéoscience… CReAAH    EA,EA,UMR6566   
#>  8 Histoire, Archéologie et Littératures des mondes … CIHAM     UMR5648         
#>  9 Centre Camille Jullian - Histoire et archéologie … CCJ       UMR7299,UMR7299…
#> 10 Archéologie des Amériques                          ArchAm    UMR8096         
#> 11 Archéologie et Archéométrie                        ArAr      UMR5138         
#> 12 Archéologie des Sociétés Méditerranéennes          ASM       UMR5140         
#> 13 Archéologie, Terre, Histoire, Sociétés [Dijon]     ARTeHiS   UMR6298         
#> 14 Archéologie et histoire ancienne : Méditerranée -… ARCHIMEDE UMR7044         
#> 15 Archéologie et Philologie d'Orient et d'Occident   AOROC     UMR8546         

# }