Facet Search

Facet Search

hal_facet(x, ...)

# S3 method for HALQuery
hal_facet(
  x,
  field,
  limit = 5,
  sort = c("index", "count"),
  prefix = NULL,
  contains = NULL,
  pivot = NULL,
  range = NULL,
  ignore_case = FALSE,
  ...
)

Arguments

x
An object of class HALQuery (typically returned by hal_api).

...
Currently not used.

field
A character string specifying the field to group by.

limit
An integer giving the maximum number of results per group.

sort
A character string specifying the field to be used to sort the results.

prefix
A character string specifying

contains
A character string specifying

pivot
A character string specifying

range
A list containing the following components: "range", "start", "end", "gap".

ignore_case
A logical scalar: should character case be ignored?

Value

An object of class HALQuery.

See also

Author

N. Frerebeau

Examples

# \donttest{
library(magrittr)

## Most recent publication by journal
hal_api() %>%
  hal_query("archéologie") %>%
  hal_facet(
    field = "docType_s",
    limit = 100,
    sort = "count",
    range = list(
      range = "producedDateY_i",
      start = 2000,
      end = 2020,
      gap = 1
    )
  ) %>%
  hal_search(limit = 10)

#> 

#> $facet_queries
#> NULL
#> 
#> $facet_fields
#> $facet_fields$docType_s
#> # A tibble: 40 x 2
#>    term   value
#>    <chr>  <chr>
#>  1 ART    14009
#>  2 IMG    13436
#>  3 COUV   9598 
#>  4 COMM   8518 
#>  5 REPORT 2399 
#>  6 OTHER  2356 
#>  7 DOUV   1651 
#>  8 OUV    1576 
#>  9 POSTER 604  
#> 10 THESE  487  
#> # … with 30 more rows
#> 
#> 
#> $facet_pivot
#> NULL
#> 
#> $facet_dates
#> NULL
#> 
#> $facet_ranges
#> $facet_ranges$producedDateY_i
#> # A tibble: 20 x 2
#>    term  value
#>    <chr> <chr>
#>  1 2000  417  
#>  2 2001  430  
#>  3 2002  525  
#>  4 2003  676  
#>  5 2004  805  
#>  6 2005  1106 
#>  7 2006  1341 
#>  8 2007  1411 
#>  9 2008  1597 
#> 10 2009  1676 
#> 11 2010  1851 
#> 12 2011  2218 
#> 13 2012  2611 
#> 14 2013  2890 
#> 15 2014  3084 
#> 16 2015  2988 
#> 17 2016  3313 
#> 18 2017  3111 
#> 19 2018  3306 
#> 20 2019  2889 
#> 
#> 

# }