Search

Search

hal_count(x, ...)

hal_search(x, ...)

# S3 method for HALQuery
hal_count(x, path = "search", ...)

# S3 method for HALQuery
hal_search(
  x,
  path = "search",
  limit = 30,
  start = 0,
  cursor = FALSE,
  parse = "df",
  concat = ",",
  progress = getOption("odyssey.progress"),
  verbose = getOption("odyssey.verbose"),
  ...
)

Arguments

x
An object of class HALQuery (typically returned by hal_api).

...
Currently not used.

path
A character string specifying the url path (see details).

limit
An integer giving the maximum number of results. According to HAL policy, it cannot exceed 10000.

start
An integer specifying an absolute offset in the complete sorted list of matches to be used as the beginning of the current page. Only used if cursor is FALSE.

cursor
A logical scalar: should a cursor be used for the pagination of results? If TRUE, the sort parameter of the query will set to "docid asc".

parse
A character string specifying the type of the results. It must be one of "df" (default) or "list" (see SolrClient).

concat
A character string specifying the character to concatenate elements of longer than length \(1\).

progress
A logical scalar: should a progress bar for for the request be printed? Only used if R is being used interactively.

verbose
A logical scalar: should extra information be reported?

Value

hal_count returns the total number of results.

hal_search returns a data.frame or list (according to type).

References

Apache Solr documentation.

HAL search documentation.

HAL reference frame documentation.

See also

Author

N. Frerebeau

Examples

# \donttest{
library(magrittr)

## Simple search
topic <- list("archéologie", "Celtes", "France") # Combined with AND
hal_api() %>%
  hal_query(topic) %>%
  hal_search(limit = 10)

#> 10 documents out of a maximum of 78 were returned.

#> # A tibble: 10 × 2
#>      docid label_s                                                              
#>      <int> <chr>                                                                
#>  1  106209 Germaine Leman-Delerive, Eugène Warmembol. La Belgique et le Nord de…
#>  2  106235 Germaine Leman-Delerive. La Belgique et le Nord de la France du IIIè…
#>  3  818910 Thierry Janin, Michel Py. Grecs et Celtes en Languedoc. Hermary A. e…
#>  4 2031928 Rapport d’évaluation d’une entité de recherche. AOROC - Archéologie …
#>  5 1094653 Thierry Lejars. Le cimetière de la Fosse-Cotheret à Roissy-en-France…
#>  6  962465 Stephane Verger. Trésors de femmes en France centrale et méridionale…
#>  7 1094643 Thierry Lejars. Images celtiques du Pays de France. Gaulois d'ici & …
#>  8  962398 Stephane Verger. Les femmes du Premier âge du Fer en France centrale…
#>  9  963605 Stephane Verger. Avant propos. Sophie Krausz, Anne Colin, Katherine …
#> 10 1536365 Marion Saurel. De la vaisselle pour les défunts en Champagne à l’âge…

## Get a list of archaeological journals
topic <- c("archéologie", "archaeology", "archäologie") # Combined with OR
hal_api() %>%
  hal_query(topic) %>%
  hal_select("title_s", "issn_s") %>%
  hal_filter("issn_s" %IN% "" %TO% "*") %>%
  hal_sort("title_s") %>%
  hal_search(path = "ref/journal")

#> 30 documents out of a maximum of 591 were returned.

#> # A tibble: 30 × 2
#>    title_s                                                              issn_s  
#>    <chr>                                                                <chr>   
#>  1 "15 ans d'archéologie en Provence-Alpes-Côte d'Azur"                 NA      
#>  2 "2020 IMEKO TC-4 International Conference on Metrology for Archaeol… NA      
#>  3 "2e Colloque international sur l'Archéologie de Paros et des Cyclad… NA      
#>  4 "ACUA. Underwater Archaeology Proceedings"                           1074-34…
#>  5 "AIMA Newsletter"                                                    1446-89…
#>  6 "ARCHÉOLOGIE DU SEL : TECHNIQUES ET SOCIÉTÉS"                        NA      
#>  7 "ATHAR\" Revue scientifique d'archéologie et du patrimoine publiée … NA      
#>  8 "Abenteuer Archäologie : Kulturen-Menschen-Monumente"                NA      
#>  9 "Abstracts of the 69th Annual Meeting Society for American Archaeol… NA      
#> 10 "Acta Archaeologica Carpathica"                                      0001-52…
#> # … with 20 more rows

## Get the most recent archaeological publication (in French) by journal
hal_api() %>%
  hal_query("archéologie") %>%
  hal_select("producedDate_tdate") %>%
  hal_filter("docType_s" %IN% "ART") %>%
  hal_sort("producedDate_tdate", decreasing = TRUE) %>%
  hal_group(
    field = "journalTitle_s",
    sort = "producedDate_tdate",
    decreasing = TRUE
  ) %>%
  hal_search(limit = 10)

#> 

#>                                                                          groupValue
#> 1                                                          Quaternary Geochronology
#> 2                                                             Nature Communications
#> 3                                           International Journal of Paleopathology
#> 4                                                                Scientific Reports
#> 5                                                        L'information grammaticale
#> 6                                                                      Archaeometry
#> 7  Annuaire de l’École pratique des hautes études. Section des sciences religieuses
#> 8                                                   Revue archéologique de Picardie
#> 9                                                                Métropolitiques.eu
#> 10                                    Journal of Data Mining and Digital Humanities
#>    numFound start   producedDate_tdate
#> 1        30     0 2022-01-01T00:00:00Z
#> 2         3     0 2021-12-01T00:00:00Z
#> 3         6     0 2021-12-01T00:00:00Z
#> 4        20     0 2021-12-01T00:00:00Z
#> 5         1     0 2021-10-01T00:00:00Z
#> 6        35     0 2021-10-01T00:00:00Z
#> 7        17     0 2021-10-01T00:00:00Z
#> 8        58     0 2021-10-01T00:00:00Z
#> 9         1     0 2021-09-30T00:00:00Z
#> 10        4     0 2021-09-28T00:00:00Z

## Get a list of archaeological laboratories
## (only joint laboratories of the CNRS and a French university)
topic <- "text" %IN% "archéologie" %AND% "code_t" %IN% "UMR"
hal_api() %>%
  hal_query(topic) %>%
  hal_select("name_s", "acronym_s", "code_s") %>%
  hal_filter("valid_s" %IN% "VALID") %>%
  hal_sort("acronym_s", decreasing = TRUE) %>%
  hal_search(path = "ref/structure", limit = 15)

#> 

#> # A tibble: 0 × 0

# }